requirements.txt
# Core scraping
playwright>=1.40.0
requests>=2.31.0

# Embeddings
torch>=2.1.0
transformers>=4.36.0
Pillow>=10.1.0

# Database
# (no extra deps needed - using requests)

# Utilities
python-dotenv>=1.0.0

# For development/testing
pytest>=7.4.0
pytest-asyncio>=0.21.0